Move `attic/training/dataset.py` -> `timemachine/training/dataset.py`
