+++ attic/training/readme.md
@@ -1,6 +1,5 @@
 * `abfe.py`, `config.ini` -- script for estimating absolute binding free energy
 * `bootstrap.py` -- functinos for bootstrapping TI estimates
-* `dataset.py` -- class for batching and splitting datasets
 * `hydration_fe.py`, `hydration_cfg.ini` -- script for training to hydration free energies
   * note: also contains a function `recenter(conf, box)`
 * `hydration_model.py` -- defines a `simulate` function that computes hydration free energies and an estimate of parameter gradients
