+++ tests/test_training.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 
-from training.dataset import Dataset
+from timemachine.training.dataset import Dataset
 
 from testsystems.relative import hif2a_ligand_pair
